Extract authOptions into own file
