+++ app/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth"
 import { redirect } from 'next/navigation'
-import { authOptions } from "./api/auth/[...nextauth]/route";
+import { authOptions } from "./api/auth/[...nextauth]/options";
 
 export default async function Home() {
   const session = await getServerSession(authOptions);
